fix: Vite crash on reload by excluding mock SDK from optimization
- Added optimizeDeps.exclude for @base44/sdk in vite.config.js
- This prevents Vite's pre-bundling issues with local file packages
- PredefinedWorkouts page now loads and reloads without crashing
- Simplified PredefinedWorkouts.jsx to minimal working version
- Updated .memory with fix documentation

Specialist's diagnosis was correct: local package causing dep optimization issues
