+++ frontend/debug-vite.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+NODE_OPTIONS="--trace-uncaught --trace-warnings --unhandled-rejections=strict" npm run dev -- --host 0.0.0.0
