+++ frontend/src/pages/PredefinedWorkouts-Minimal.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export default function PredefinedWorkouts() {
+  console.log("PredefinedWorkouts page loaded");
+  
+  return (
+    <div className="space-y-6">
+      <h1 className="text-3xl font-bold text-gray-900">Predefined Workouts</h1>
+      <p className="text-lg text-gray-600">Minimal version - no data loading</p>
+    </div>
+  );
+}
